[배인수] 1. Game Current 세팅하는 로직 수정 2. 소스 다듬기

+++ Assets/Scripts/Game/GameProtocolDispatcher.cs
@@ -6,7 +6,7 @@ using System;
 
 public class GameProtocolDispatcher : MonoBehaviour
 {
-    private Dictionary<int, Action<IPhotonEventMessage>> protocolHandlers = new Dictionary<int, Action<IPhotonEventMessage>>();
+    private Dictionary<byte, Action<IPhotonEventMessage>> protocolHandlers = new Dictionary<byte, Action<IPhotonEventMessage>>();
 
     private void Awake()
     {
